+++ app/assets/javascripts/govuk_publishing_components/analytics-ga4/ga4-core.js
@@ -135,6 +135,8 @@ window.GOVUK.analyticsGa4 = window.GOVUK.analyticsGa4 || {};
               return 'command/win click'
             } else if (event.shiftKey) {
               return 'shift click'
+            } else if (event.altKey) {
+              return 'alt/option click'
             } else {
               return 'primary click'
             }